feat : 학사일정 안내 모듈 구현
- getScheduleDict 모듈로 학사일정 파일을 읽어와 딕셔너리를 생성
- index.js 는 state = 1 일 때 shedule 모듈을 실행하는 것으로 구조 변경
- schedule 모듈은 state 를 인자로 받지 않음. 날짜 체크를 index.js 가 아닌 모듈 안에서 수행
