+++ index.js
@@ -19,48 +19,22 @@ rtm.start();
 const greeting = require('./greeting');
 const square = require('./square');
 const schedule = require('./schedule'); // 학사일정 안내 모듈
-const getschedule = require('./getschedule'); // 학사일정 딕셔너리 생성 모듈. 학사일정 안내 시 매번 for문을 돌지 않도록 함.
+const getScheduleDict = require('./getScheduleDict'); // 학사일정 딕셔너리 생성 모듈. 학사일정 안내 시 매번 for문을 돌지 않도록 함.
 
-function dateExist(text) {
-  const m1 = [1, 3, 5, 7, 8, 10, 12];
-  const m2 = [4, 6, 9, 11];
-
-  const month = Number(text.split('/')[0]);
-  const day = Number(text.split('/')[1]);
-
-  if (m1.includes(month) && day >= 1 && day <= 31) {
-    return true;
-  } if (m2.includes(month) && day >= 1 && day <= 30) {
-    return true;
-  } if (month === 2 && day >= 1 && day <= 29) {
-    return true;
-  }
-  return false;
-}
 /* eslint no-restricted-globals: 0 */
-const dateRegex = /\d{1,2}\/\d{1,2}/; // 월/일
 let state = 0;
-const scheduledict = getschedule(); // 학사일정 딕셔너리 가져오기
+const scheduledict = getScheduleDict(); // 학사일정 딕셔너리 가져오기
 
 rtm.on('message', (message) => {
   const { channel } = message;
   const { text } = message;
 
-  if (!isNaN(text)) {
+  if (state === 1) { // 학사일정을 입력한 상태면 학사일정 안내 모듈 실행
+    schedule(rtm, text, channel, scheduledict);
+    state = 0; // 상태 초기화
+  } else if (!isNaN(text)) {
     square(rtm, text, channel);
-    state = 0;
-  } else if (state === 0 && dateRegex.test(text) && dateExist(text)) {
-    // 날짜 양식과 부합하나 학사일정을 먼저 입력하지 않은 경우
-    schedule(rtm, state, text, channel, scheduledict);
-  } else if (state === 1) {
-    if (!dateRegex.test(text) || !dateExist(text)) { // 날짜 양식이 틀렸거나 존재하지 않는 날짜의 경우
-      state = 2;
-    }
-    schedule(rtm, state, text, channel, scheduledict);
-    state = 0;
   } else {
-    state = 0; // 학사일정 입력 후 날짜양식을 바로 입력하지 않으면 초기화
-
     switch (text) {
       case '테스트를 시작한다.':
         break;
@@ -68,11 +42,11 @@ rtm.on('message', (message) => {
         greeting(rtm, channel);
         break;
       case '학사일정':
-        schedule(rtm, state, text, channel);
-        state = 1; // 학사일정을 먼저 입력한 경우
+        schedule(rtm, text, channel, scheduledict);
+        state = 1; // 학사일정을 입력하면 날짜를 입력받을 수 있는 state 바뀜
         break;
       default:
-        rtm.sendMessage('I`m alive', channel);
+        rtm.sendMessage('"hi" "학사일정" "오늘 밥 뭐야" 를 입력하거나 "학과이름을 영어로" 입력하시오', channel);
     }
   }
 });
